+++ core/domain/prod_validators_test.py
@@ -275,129 +275,6 @@ class UsernameChangeAuditModelValidatorTests(test_utils.AuditJobsTestBase):
             expected_output, sort=True, literal_eval=False)
 
 
-class TrainingJobExplorationMappingModelValidatorTests(
-        test_utils.AuditJobsTestBase):
-
-    def setUp(self):
-        super(TrainingJobExplorationMappingModelValidatorTests, self).setUp()
-
-        self.signup(self.OWNER_EMAIL, self.OWNER_USERNAME)
-
-        self.owner_id = self.get_user_id_from_email(self.OWNER_EMAIL)
-
-        explorations = [exp_domain.Exploration.create_default_exploration(
-            '%s' % i,
-            title='title %d' % i,
-            category='category%d' % i,
-        ) for i in python_utils.RANGE(2)]
-
-        for exp in explorations:
-            exp.add_states(['StateTest%s' % exp.id])
-            exp_services.save_new_exploration(self.owner_id, exp)
-
-        id0 = classifier_models.TrainingJobExplorationMappingModel.create(
-            '0', 1, 'StateTest0', 'job0')
-        self.model_instance_0 = (
-            classifier_models.TrainingJobExplorationMappingModel.get_by_id(id0))
-        id1 = classifier_models.TrainingJobExplorationMappingModel.create(
-            '1', 1, 'StateTest1', 'job1')
-        self.model_instance_1 = (
-            classifier_models.TrainingJobExplorationMappingModel.get_by_id(id1))
-
-        self.job_class = (
-            prod_validation_jobs_one_off
-            .TrainingJobExplorationMappingModelAuditOneOffJob)
-
-    def test_standard_operation(self):
-        expected_output = [
-            u'[u\'fully-validated TrainingJobExplorationMappingModel\', 2]']
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_created_on_greater_than_last_updated(self):
-        self.model_instance_0.created_on = (
-            self.model_instance_0.last_updated + datetime.timedelta(days=1))
-        self.model_instance_0.update_timestamps()
-        self.model_instance_0.put()
-        expected_output = [(
-            u'[u\'failed validation check for time field relation check '
-            'of TrainingJobExplorationMappingModel\', '
-            '[u\'Entity id %s: The created_on field has a value '
-            '%s which is greater than the value '
-            '%s of last_updated field\']]') % (
-                self.model_instance_0.id,
-                self.model_instance_0.created_on,
-                self.model_instance_0.last_updated
-            ), u'[u\'fully-validated TrainingJobExplorationMappingModel\', 1]']
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-    def test_model_with_last_updated_greater_than_current_time(self):
-        self.model_instance_1.delete()
-        expected_output = [(
-            u'[u\'failed validation check for current time check of '
-            'TrainingJobExplorationMappingModel\', '
-            '[u\'Entity id %s: The last_updated field has a '
-            'value %s which is greater than the time when the job was run\']]'
-        ) % (self.model_instance_0.id, self.model_instance_0.last_updated)]
-
-        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
-            hours=13)
-        with datastore_services.mock_datetime_for_datastore(mocked_datetime):
-            self.run_job_and_check_output(
-                expected_output, sort=True, literal_eval=False)
-
-    def test_missing_exploration_model_failure(self):
-        exp_models.ExplorationModel.get_by_id('0').delete(
-            feconf.SYSTEM_COMMITTER_ID, '', [])
-        expected_output = [
-            (
-                u'[u\'failed validation check for exploration_ids field '
-                'check of TrainingJobExplorationMappingModel\', '
-                '[u"Entity id %s: based on field exploration_ids having value '
-                '0, expected model ExplorationModel with id 0 but it doesn\'t '
-                'exist"]]') % self.model_instance_0.id,
-            u'[u\'fully-validated TrainingJobExplorationMappingModel\', 1]']
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-    def test_invalid_exp_version(self):
-        model_instance_with_invalid_exp_version = (
-            classifier_models.TrainingJobExplorationMappingModel(
-                id='0.5.StateTest0', exp_id='0', exp_version=5,
-                state_name='StateTest0', job_id='job_id'))
-        model_instance_with_invalid_exp_version.update_timestamps()
-        model_instance_with_invalid_exp_version.put()
-        expected_output = [
-            (
-                u'[u\'failed validation check for exp version check '
-                'of TrainingJobExplorationMappingModel\', [u\'Entity id %s: '
-                'Exploration version 5 in entity is greater than the '
-                'version 1 of exploration corresponding to exp_id 0\']]'
-            ) % model_instance_with_invalid_exp_version.id,
-            u'[u\'fully-validated TrainingJobExplorationMappingModel\', 2]']
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-    def test_invalid_state_name(self):
-        model_instance_with_invalid_state_name = (
-            classifier_models.TrainingJobExplorationMappingModel(
-                id='0.1.invalid', exp_id='0', exp_version=1,
-                state_name='invalid', job_id='job_id'))
-        model_instance_with_invalid_state_name.update_timestamps()
-        model_instance_with_invalid_state_name.put()
-        expected_output = [
-            (
-                u'[u\'failed validation check for state name check '
-                'of TrainingJobExplorationMappingModel\', [u\'Entity id %s: '
-                'State name invalid in entity is not present in '
-                'states of exploration corresponding to exp_id 0\']]'
-            ) % model_instance_with_invalid_state_name.id,
-            u'[u\'fully-validated TrainingJobExplorationMappingModel\', 2]']
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-
 class CollectionModelValidatorTests(test_utils.AuditJobsTestBase):
 
     def setUp(self):
